Use a loop to populate cms
As we loop on the size of the configmap template search we populate the
cms slice with pointers to the different configmap structs.  With the
interator we point to the same part of the iterator multiple times,
effectivly pointing to the last item to be iterated on multiple times.

Issue #3